refactor: refactor files to atomic architecture
